Fix fetching of comments
