+++ frontend/components/dining-hall/DiningHall.js
@@ -1,4 +1,4 @@
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import { Keyboard, Text, SafeAreaView, TouchableWithoutFeedback } from 'react-native';
 
 import diningHallStyles from './DiningHallStyles';
@@ -15,9 +15,11 @@ import { DINING_HALLS } from '../../constants';
 const DiningHall = ({ navigation, route }) => {
     const [isMenuVisible, setIsMenuVisible] = useState(false);
     const [isRatingInputVisible, setIsRatingInputVisible] = useState(false);
-    const { id } = route.params; // TODO: use id to fetch dining hall data
+    const [comments, setComments] = useState([]);
 
-    // Get dining hall name from id
+    const { id } = route.params;
+
+    // TODO: fetch name from backend
     const diningHallName = DINING_HALLS[id].name
 
     const closeKeyboardAndMenu = () => {
@@ -42,6 +44,32 @@ const DiningHall = ({ navigation, route }) => {
         setIsRatingInputVisible(false);
     }
 
+    const fetchComments = () => {
+        const requestOptions = {
+            method: 'GET',
+            headers: {'Content-Type': 'application/json'}
+        };
+
+        fetch(`https://sleepy-reaches-22563.herokuapp.com/api/getAll/diningComments/${id}`, requestOptions)
+        .then(async response => {
+            const isJson = response.headers.get('content-type')?.includes('application/json');
+            const data = isJson && await response.json();
+
+            if (!response.ok) {
+                const err = (data && data.message) || response.status;
+                return Promise.reject(err);
+            }
+            setComments(data);
+            
+            console.log("Successfully fetched comments.");
+        }).catch(error => {
+            console.log(error);
+            alert("Comments failed to load. Please try again later.");
+        });
+    }
+
+    useEffect(fetchComments, []);
+
     // TODO: use dining hall data to render dining hall info
     return (
         <TouchableWithoutFeedback onPress={closeKeyboardAndMenu} accessible={false}>
@@ -49,8 +77,8 @@ const DiningHall = ({ navigation, route }) => {
                 <Text style={diningHallStyles.diningHallName}>{diningHallName}</Text>
                 <OverallRating diningHallId={id} onPress={handleOverallRatingPress} />
                 <Subratings diningHallId={id} />
-                <Comments diningHallId={id} />
-                <CommentInput diningHallId={id} />
+                <Comments comments={comments} />
+                <CommentInput diningHallId={id} onSubmit={fetchComments} />
                 <BackButton onPress={handleBackButtonPress} />
                 <DotsMenu diningHallId={id} isMenuVisible={isMenuVisible} toggleMenu={handleDotsMenuPress} />
                 {isRatingInputVisible && <RatingInput diningHallId={id} onClose={handleRatingInputClose} />}
